+++ cpp/src/dataframe/column.cpp
@@ -200,7 +200,7 @@ namespace{
 gdf_size_type gdf_dtype_size(gdf_dtype dtype) {
   return cudf::type_dispatcher(dtype, get_type_size{});
 }
-
+ 
 // Returns the size in bytes of the data type of the gdf_column
 gdf_error get_column_byte_width(gdf_column * col, 
                                 int * width)