+++ python/cudf/dataframe/dataframe.py
@@ -623,7 +623,7 @@ class DataFrame(object):
         l_old_cols = len(old_cols)
         l_new_cols = len(columns)
         if l_new_cols != l_old_cols:
-            msg = f'Length of new column names: {l_new_cols} does not ' \
+            msg = 'Length of new column names: {l_new_cols} does not ' \
                   'match length of previous column names: {l_old_cols}'
             raise ValueError(msg)
 
@@ -642,7 +642,7 @@ class DataFrame(object):
         old_length = len(self._index)
 
         if new_length != old_length:
-            msg = f'Length mismatch: Expected index has {old_length}' \
+            msg = 'Length mismatch: Expected index has {old_length}' \
                     ' elements, new values have {new_length} elements'
             raise ValueError(msg)